Added converter to\from XML entities

+++ WHPP-MobileEntity/src/main/java/org/courses/mobileentity/entity/CoordsXML.java
@@ -25,9 +25,9 @@ public class CoordsXML implements Serializable {
 
 	private String address;
 
-	private int latitude;
+	private double latitude;
 
-	private int longitude;
+	private double longitude;
 
 	private List<RoutepointXML> routepointList;
 
@@ -64,19 +64,19 @@ public class CoordsXML implements Serializable {
 		this.address = address;
 	}
 
-	public int getLatitude() {
+	public double getLatitude() {
 		return latitude;
 	}
 
-	public void setLatitude(int latitude) {
+	public void setLatitude(double latitude) {
 		this.latitude = latitude;
 	}
 
-	public int getLongitude() {
+	public double getLongitude() {
 		return longitude;
 	}
 
-	public void setLongitude(int longitude) {
+	public void setLongitude(double longitude) {
 		this.longitude = longitude;
 	}
